Moved master commands to own function, introduction of Twitch API, new base commands 'GAME' and 'STATUS', allows bot to send requests to twitch to get/set game/status data. Updated web channels page to include commands & quotes list, username links to their twitch page. web/api/globals* for future use later. New configuration file for web, used atm for channel configs locations.

+++ web/channels_commands.php
@@ -1,14 +1,19 @@
 <?PHP
-	$fil = file_get_contents("../cfg/#" . $_GET['c'] . "/cmds.cfg");
-	$commands = explode("\r\n", $fil);
-	
-	echo "<br /><p>Command list for " . ucwords(strtolower($_GET['c'])) . "</p>";
-	echo "<table class=\"minTable\"><tr><th>Trigger</th><th>Rank</th><th>Active</th><th>Text</th></tr>";
-	foreach ($commands as $command) {
-		if($command !== "") {
-			$command_segments = explode("|", $command, 4);
-			echo "<tr><td>" . $command_segments[2] . "</td><td>" . $command_segments[0] . "</td><td>" . $command_segments[1] . "</td><td> " . $command_segments[3] . "</td></tr>";
+	$filPath = "../cfg/#" . $_GET['c'] . "/cmds.cfg";
+	if(file_exists($filPath)) {
+		$fil = file_get_contents($filPath);
+		$commands = explode("\r\n", $fil);
+		
+		echo "<br /><p>Command list for " . ucwords(strtolower($_GET['c'])) . "</p>";
+		echo "<table class=\"minTable\"><tr><th>Trigger</th><th>Rank</th><th>Active</th><th>Text</th></tr>";
+		foreach ($commands as $command) {
+			if($command !== "") {
+				$command_segments = explode("|", $command, 4);
+				echo "<tr><td>" . $command_segments[2] . "</td><td>" . $command_segments[0] . "</td><td>" . $command_segments[1] . "</td><td> " . $command_segments[3] . "</td></tr>";
+			}
 		}
+		echo"</table>";
+	} else {
+		require("error.php");
 	}
-	echo"</table>";
 ?>
